+++ rs/moq-uring/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.4](https://github.com/moq-dev/moq/compare/moq-uring-v0.0.3...moq-uring-v0.0.4) - 2026-09-25
+
+### Added
+
+- *(moq-uring)* report a session's peer address and SNI to auth ([#4056](https://github.com/moq-dev/moq/pull/4056))
+
+## [0.0.3](https://github.com/moq-dev/moq/compare/moq-uring-v0.0.2...moq-uring-v0.0.3) - 2026-09-24
+
+### Other
+
+- updated the following local packages: moq-net
+
 ## [0.0.2](https://github.com/moq-dev/moq/compare/moq-uring-v0.0.1...moq-uring-v0.0.2) - 2026-09-23
 
 ### Fixed